SCC: extract content.py file classes into files
