+++ src/test/python/test_scc_paragraph.py
@@ -31,9 +31,9 @@ import unittest
 from fractions import Fraction
 
 from ttconv.model import ContentDocument, Span, Br
-from ttconv.scc.content import SccCaptionLine
 from ttconv.scc.paragraph import SccCaptionParagraph
 from ttconv.scc.style import SccCaptionStyle
+from ttconv.scc.caption_line import SccCaptionLine
 from ttconv.style_properties import TextAlignType
 from ttconv.time_code import SmpteTimeCode, FPS_30
